rename llvm2 -> llvm
