+++ weld/src/codegen/llvm/serde.rs
@@ -11,7 +11,7 @@ use std::ffi::CStr;
 
 use crate::ast::Type::*;
 use crate::ast::*;
-use crate::codegen::llvm2::vector::VectorExt;
+use crate::codegen::llvm::vector::VectorExt;
 use crate::error::*;
 use crate::sir::*;
 
@@ -528,7 +528,7 @@ impl DeHelper for LlvmGenerator {
         buffer: LLVMValueRef,
         position: LLVMValueRef,
     ) -> WeldResult<(LLVMValueRef, LLVMValueRef)> {
-        use crate::codegen::llvm2::vector::VectorExt;
+        use crate::codegen::llvm::vector::VectorExt;
 
         let size = self.size_of(ty);
         let pointer = self.gen_at(builder, &SER_TY, buffer, position)?;
@@ -554,7 +554,7 @@ impl DeHelper for LlvmGenerator {
         buffer: LLVMValueRef,
         position: LLVMValueRef,
     ) -> WeldResult<LLVMValueRef> {
-        use crate::codegen::llvm2::vector::VectorExt;
+        use crate::codegen::llvm::vector::VectorExt;
         let elem_size = self.size_of(LLVMGetElementType(LLVMTypeOf(ptr)));
         let size = LLVMBuildNSWMul(builder, size, elem_size, c_str!(""));
 
@@ -741,7 +741,7 @@ impl DeHelper for LlvmGenerator {
                     // NOTE: This requires re-hashing: we could look into encoding dictionaries without
                     // having to do this.
                     use self::llvm_sys::LLVMIntPredicate::LLVMIntSGT;
-                    use crate::codegen::llvm2::hash::GenHash;
+                    use crate::codegen::llvm::hash::GenHash;
 
                     let size_type = self.i64_type();
                     let (size, start_position) =
